pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["poc_homography"]

[tool.hatch.build.targets.wheel.sources]
"poc_homography" = "poc_homography"

[tool.hatch.build]
include = [
    "poc_homography/**/*.py",
    "poc_homography/**/*.j2",
]

[project]
name = "poc-homography"
version = "0.1.0"
description = "POC for camera homography and coordinate transformations"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "django>=5.0",
    "jinja2>=3.1.6",
    "numpy>=1.24",
    "opencv-python>=4.7.0",
    "pyproj>=3.7.1",
    "PyYAML>=6.0",
    "requests>=2.28",
    "scipy>=1.9",
    "typer>=0.9.0",
]

[project.scripts]
hom = "poc_homography.cli.main:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.100",
    "pyright>=1.1.390",
    "ruff>=0.4",
    "pre-commit>=3.5",
    "types-PyYAML",
    "types-requests",
    "matplotlib>=3.7",
    "vulture>=2.14",
    "deadcode>=2.4",
    "poethepoet>=0.31",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["poc_homography", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B007",   # loop control variable not used (common pattern)
    "B008",   # do not perform function calls in argument defaults
    "B028",   # no explicit stacklevel in warnings
    "B904",   # raise from err (stylistic)
    "B905",   # zip without strict (Python 3.10+ style)
    "C901",   # too complex
    "ARG001", # unused function argument (common in callbacks/interfaces)
    "ARG002", # unused method argument (common in callbacks/interfaces)
    "F821",   # undefined name (forward references in type hints)
    "F841",   # local variable assigned but never used (cleanup later)
    "PTH",    # pathlib suggestions (not critical, keep existing os.path usage)
    "RUF001", # ambiguous unicode in strings (intentional math symbols)
    "RUF002", # ambiguous unicode in docstrings (math symbols)
    "RUF003", # ambiguous unicode in comments (math symbols)
    "RUF005", # list concatenation vs spread (stylistic)
    "RUF012", # mutable class attributes (common pattern)
    "RUF013", # implicit Optional (legacy code)
    "RUF022", # unsorted __all__ (preserve logical grouping)
    "RUF043", # regex pattern metacharacters (intentional)
    "RUF059", # unused unpacked variable (use _ prefix manually)
    "SIM102", # nested if statements (sometimes clearer)
    "SIM103", # return condition directly (sometimes less clear)
    "SIM105", # contextlib.suppress suggestion (stylistic)
    "SIM108", # ternary operator (sometimes less readable)
    "SIM115", # context manager for files (not always applicable)
    "SIM118", # key in dict vs dict.keys() (stylistic)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "B011", "E402", "S101", "SIM117"]  # Tests have asserts and path setup

[tool.ruff.lint.isort]
known-first-party = ["poc_homography"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"
include = ["poc_homography", "tests", "examples"]
exclude = ["build", "dist", ".venv", "**/__pycache__", "main.py", "ptz_discovery_and_control"]
extraPaths = ["."]
reportMissingImports = true
reportMissingTypeStubs = false

# Per-directory overrides: relaxed type checking for tests
# Production code (poc_homography/) uses default strict NewType checking
# Tests use relaxed checking since they intentionally pass raw literals
[[tool.pyright.executionEnvironments]]
root = "tests"
reportArgumentType = "none"
reportOptionalSubscript = "none"
reportOptionalMemberAccess = "none"
reportOptionalIterable = "none"
reportCallIssue = "none"
reportOperatorIssue = "none"
reportReturnType = "none"
reportAbstractUsage = "none"
reportMissingImports = "none"
reportUndefinedVariable = "none"

[[tool.pyright.executionEnvironments]]
root = "examples"
reportArgumentType = "none"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["poc_homography"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[tool.vulture]
min_confidence = 80
paths = ["poc_homography"]
exclude = ["tests/", "examples/", "build/", "dist/"]
# Ignore interface parameters required by ABC but not used in all implementations
ignore_names = ["frame", "kwargs"]

[tool.deadcode]
exclude = ["build", "dist"]

[tool.poe.tasks]
lint = "ruff check poc_homography"
lint-fix = "ruff check poc_homography --fix"
format = "ruff format poc_homography"
format-check = "ruff format --check poc_homography"
typecheck = "pyright poc_homography"
vulture = "vulture poc_homography --min-confidence 80"
deadcode = "deadcode . --only poc_homography/"
deadcode-fix = "deadcode . --only poc_homography/ --fix"
test = "pytest tests/ -v"
test-cov = "pytest tests/ -v --cov=poc_homography --cov-report=term-missing"
pre-commit = "pre-commit run --all-files"

[tool.poe.tasks.validate]
sequence = ["lint", "format-check", "typecheck", "vulture"]
help = "Run all validation checks (no tests)"

[tool.poe.tasks.ci]
sequence = ["lint", "format-check", "typecheck", "vulture", "test"]
help = "Run full CI: validation + tests"

[tool.poe.tasks.clean]
shell = "rm -rf build/ dist/ *.egg-info .pytest_cache .coverage htmlcov/ && find . -type d -name __pycache__ -exec rm -rf {} + 2>/dev/null || true"
help = "Clean build artifacts"
